+++ RELEASE_NOTES.md
@@ -14,6 +14,7 @@
 
 - Hourly and daily CPU graph now available
 - Host missing time calculation fixed
+- Javascript split into its own source file